intial commit

+++ app/Controllers/AccountController.php
@@ -12,22 +12,17 @@ class AccountController extends ResourceController
 {
     protected $modelName = 'App\Models\AccountModel';
     protected $format    = 'json';
+    protected $allowedColumns = [];
 
     public function index()
     {
-
         $params = $this->request->getVar(['columns', 'sort', 'page', 'pageSize']);
-        $allowedColumns = [];
-
-        $response = new ApiResponse($this->model, $params, $allowedColumns);
-
-        return $response->getCollectionResponse();
+        $response = new ApiResponse($this->model, $params, $this->allowedColumns);
+        return $response->getCollectionResponse(true, ['owner', 'orgid']);
     }
 
     public function create()
     {
-        $data = (array)$this->request->getVar();
-
         $rules = config('Validation')->create['accounts'];
         // Validate input
         if (!$this->validate($rules)) {
@@ -37,7 +32,7 @@ class AccountController extends ResourceController
                 'error'   => $this->validator->getErrors()
             ], Response::HTTP_BAD_REQUEST);
         }
-
+        $data = $this->validator->getValidated();
         $orgModel = new OrganizationModel();
 
         if (!$orgModel->find($data['orgid'])) {
@@ -47,6 +42,9 @@ class AccountController extends ResourceController
                 'error'   =>    ["orgid" => "The Organization doesn't exist."]
             ], Response::HTTP_BAD_REQUEST);
         }
+
+        $data['owner'] = auth()->user()->owner;
+        $data['creator'] = auth()->user_id();
 
         if ($this->model->save($data)) {
             return $this->respondCreated([
@@ -65,9 +63,20 @@ class AccountController extends ResourceController
 
     public function update($id = null)
     {
-        $data = $this->request->getRawInput();
+        $account = $this->model->find($id);
+        if (!$account) {
+            return $this->respond([
+                'status'  => false,
+                'message' => "Account doesn't exist.",
+            ], Response::HTTP_NOT_FOUND);
+        }
+
+        $response = auth()->can('update', 'accounts', ['owner', 'orgid'], [$account]);
+        if ($response->denied())
+            return $response->responsed();
 
         $rules = config('Validation')->update['accounts'];
+
         // Validate input
         if (!$this->validate($rules)) {
             return $this->respond([
@@ -82,6 +91,7 @@ class AccountController extends ResourceController
                 'message' => 'Account not found'
             ], Response::HTTP_NOT_FOUND);
         }
+        $data = $this->validator->getValidated();
 
         $this->model->update($id, $data);
         return $this->respond([
@@ -93,12 +103,18 @@ class AccountController extends ResourceController
 
     public function show($id = null)
     {
+        $account = $this->model->find($id);
+        if (!$account) {
+            return $this->respond([
+                'status'  => false,
+                'message' => "Account doesn't exist.",
+            ], Response::HTTP_NOT_FOUND);
+        }
+        
         $params = $this->request->getVar(['columns', 'sort', 'page', 'pageSize']);
-        $allowedColumns = [];
-        $this->model->find($id);
-        $response = new ApiResponse($this->model, $params, $allowedColumns);
+        $response = new ApiResponse($this->model, $params, $this->allowedColumns);
 
-        return $response->getSingleResponse();
+        return $response->getSingleResponse(true, ['owner', 'orgid']);
     }
 
     public function delete($id = null)
@@ -110,6 +126,10 @@ class AccountController extends ResourceController
                 'message' => 'Account not found'
             ], Response::HTTP_NOT_FOUND);
         }
+
+        $response = auth()->can('delete', 'accounts', ['owner', 'orgid'], [$account]);
+        if ($response->denied())
+            return $response->responsed();
 
         if ($this->model->delete($id)) {
             return $this->respondDeleted([
@@ -125,52 +145,52 @@ class AccountController extends ResourceController
         }
     }
 
-     // Pull changes from the server
-     public function pull()
-     {
-         $lastSyncTime = $this->request->getGet('lastSyncTime');
- 
-         // Fetch updated after the last pulled timestamp
-         $records = $this->model
-             ->where('updated_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
-             ->findAll();
-         $deletedRecords = $this->model->select(['id', 'deleted_at'])
-             ->where('deleted_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
-             ->onlyDeleted()->findAll();
- 
-         return $this->respond([
-             'updated' => $records,
-             'deleted' => $deletedRecords,
-             'timestamp' =>  date('Y-m-d H:i:s', strtotime('now')) // Current server time for synchronization
-         ]);
-     }
- 
-     // Push changes to the server
-     public function push()
-     {
-         $rules = config('Validation')->sync;
-         // Validate input
-         if (!$this->validate($rules)) {
-             return $this->respond([
-                 'status'  => false,
-                 'message' => 'Failed validating data',
-                 'error'   => $this->validator->getErrors()
-             ], Response::HTTP_BAD_REQUEST);
-         }
- 
-         $updates = $this->request->getVar('updated');
-         $nrowsUpdated = sizeof($updates);
-         $deleted = $this->request->getVar('deleted');
-         $nrowsDeleted = sizeof($updates);
- 
-         if ($nrowsUpdated > 0)
-             $this->model->builder()->updateBatch($updates, ['id'], sizeof($updates));
-         if ($nrowsDeleted > 0)
-             $this->model->builder()->updateBatch($deleted, ['id'], sizeof($deleted));
- 
-         return $this->respond([
-             'status' => true,
-             'message' => 'Sync completed successfully'
-         ], Response::HTTP_OK);
-     }
+    // Pull changes from the server
+    public function pull()
+    {
+        $lastSyncTime = $this->request->getGet('lastSyncTime');
+
+        // Fetch updated after the last pulled timestamp
+        $records = $this->model
+            ->where('updated_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
+            ->findAll();
+        $deletedRecords = $this->model->select(['id', 'deleted_at'])
+            ->where('deleted_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
+            ->onlyDeleted()->findAll();
+
+        return $this->respond([
+            'updated' => $records,
+            'deleted' => $deletedRecords,
+            'timestamp' =>  date('Y-m-d H:i:s', strtotime('now')) // Current server time for synchronization
+        ]);
+    }
+
+    // Push changes to the server
+    public function push()
+    {
+        $rules = config('Validation')->sync;
+        // Validate input
+        if (!$this->validate($rules)) {
+            return $this->respond([
+                'status'  => false,
+                'message' => 'Failed validating data',
+                'error'   => $this->validator->getErrors()
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $updates = $this->request->getVar('updated');
+        $nrowsUpdated = sizeof($updates);
+        $deleted = $this->request->getVar('deleted');
+        $nrowsDeleted = sizeof($updates);
+
+        if ($nrowsUpdated > 0)
+            $this->model->builder()->updateBatch($updates, ['id'], sizeof($updates));
+        if ($nrowsDeleted > 0)
+            $this->model->builder()->updateBatch($deleted, ['id'], sizeof($deleted));
+
+        return $this->respond([
+            'status' => true,
+            'message' => 'Sync completed successfully'
+        ], Response::HTTP_OK);
+    }
 }
